// apps/web/fsd/shared/lib/lightbox/types.ts
export type LightboxExternalProps = DeepPartial<
  DeepPartial<
    DeepPartial<LightboxProps, 'animation' | 'toolbar' | 'noScroll'>,
    'carousel',
    'imageProps'
  >,
  'controller',
  'ref'
>;

/** Lightbox properties */
export interface LightboxProps {
  /** if `true`, the lightbox is open */
  open: boolean;
  /** a callback to close the lightbox */
  close: Callback;
  /** starting slide index */
  index: number;
  /** slides to display in the lightbox */
  slides: Slide[];
  /** custom render functions */
  render: Render;
  /** custom UI labels / translations */
  labels: Labels;
  /** enabled plugins */
  plugins: Plugin[];
  /** toolbar settings */
  toolbar: ToolbarSettings;
  /** carousel settings */
  carousel: CarouselSettings;
  /** animation settings */
  animation: AnimationSettings;
  /** controller settings */
  controller: ControllerSettings;
  /** portal settings */
  portal: PortalSettings;
  /** NoScroll module settings */
  noScroll: NoScrollSettings;
  /** lifecycle callbacks */
  on: Callbacks;
  /** customization styles */
  styles: SlotStyles;
  /** CSS class of the lightbox root element */
  className: string;
}

/** Slide */
export type Slide = SlideTypes[SlideTypeKey];
// ...
export type Component = React.ComponentType<LightBoxComponentProps>;

/** Lightbox component properties */
export type LightBoxComponentProps = React.PropsWithChildren<
  Omit<LightboxProps, 'slides' | 'index' | 'plugins'>
>;

/** Lightbox component tree node */
export type Node = {
  /** module */
  module: Module;
  /** module child nodes */
  children?: Node[];
};

/** Lightbox module */
export type Module = {
  /** module name */
  name: string;
  /** module component */
  component: Component;
};
// ...
export type SlotStyles = {
  [key in Slot]?: SlotCSSProperties;
};

/** Carousel settings */
export interface CarouselSettings {
  /** if `true`, the lightbox carousel doesn't wrap around */
  finite: boolean;
  /** the lightbox preloads (2 * preload + 1) slides */
  preload: number;
  /** padding around each slide (e.g., "16px", "10%" or 0) */
  padding: LengthOrPercentage;
  /** spacing between slides (e.g., "100px", "50%" or 0) */
  spacing: LengthOrPercentage;
  /** `object-fit` setting for image slides */
  imageFit: ImageFit;
  /** custom image attributes */
  imageProps: Omit<
    React.ImgHTMLAttributes<HTMLImageElement>,
    'src' | 'alt' | 'sizes' | 'srcSet' | 'onLoad' | 'onError' | 'onClick'
  >;
}

export type LengthOrPercentage = `${number}px` | `${number}%` | number;

/** Animation settings */
export interface AnimationSettings {
  /** fade-in / fade-out animation settings */
  fade: number;
  /** swipe animation settings */
  swipe: number;
  /** override for `swipe` animation settings when using keyboard navigation or navigation buttons */
  navigation?: number;
  /** animation timing function settings */
  easing: {
    /** fade-in / fade-out animation timing function */
    fade: string;
    /** slide swipe animation timing function */
    swipe: string;
    /** slide navigation animation timing function (when using keyboard navigation or navigation buttons) */
    navigation: string;
  };
}

/** controller settings */
export interface ControllerSettings {
  /** controller ref */
  ref: React.ForwardedRef<ControllerRef>;
  // TODO v4: remove
  /** @deprecated for internal use only */
  focus: boolean;
  // TODO v4: remove
  /** @deprecated for internal use only */
  touchAction: 'none' | 'pan-y';
  /** if `true`, set ARIA attributes on the controller div */
  aria: boolean;
  /** if `true`, close the lightbox on pull-up gesture */
  closeOnPullUp: boolean;
  /** if `true`, close the lightbox on pull-down gesture */
  closeOnPullDown: boolean;
  /** if `true`, close the lightbox when the backdrop is clicked */
  closeOnBackdropClick: boolean;
  /** if `true`, prevent default for horizontal wheel scroll stores (for internal use only) */
  preventDefaultWheelX: boolean;
  /** if `true`, prevent default for vertical wheel scroll stores (for internal use only) */
  preventDefaultWheelY: boolean;
  /** if `true`, disable slide change on pointer swipe / drag */
  disableSwipeNavigation: boolean;
}
// ...
export interface PortalSettings {
  /** portal mount point */
  root?: DocumentFragment | Element | null;
}

/** NoScroll module settings */
export interface NoScrollSettings {
  /** if `true`, the NoScroll module functionality is disabled */
  disabled: boolean;
}
// ...
export interface Render {
  /** render custom slide type, or override the default image slide */
  slide?: RenderFunction<RenderSlideProps>;
  /** render custom slide header (use absolute positioning) */
  slideHeader?: RenderFunction<RenderSlideHeaderProps>;
  /** render custom slide footer (use absolute positioning) */
  slideFooter?: RenderFunction<RenderSlideFooterProps>;
  /** render custom slide container */
  slideContainer?: RenderFunction<RenderSlideContainerProps>;
  /** render custom controls or additional elements in the lightbox (use absolute positioning) */
  controls?: RenderFunction;
  /** render custom Prev icon */
  iconPrev?: RenderFunction;
  /** render custom Next icon */
  iconNext?: RenderFunction;
  /** render custom Close icon */
  iconClose?: RenderFunction;
  /** render custom Loading icon */
  iconLoading?: RenderFunction;
  /** render custom Error icon */
  iconError?: RenderFunction;
  /** render custom Prev button */
  buttonPrev?: RenderFunction;
  /** render custom Next button */
  buttonNext?: RenderFunction;
  /** render custom Close button */
  buttonClose?: RenderFunction;
}

export type Callback<T = void> = [T] extends [void] ? () => void : (props: T) => void;

export interface ViewCallbackProps {
  index: number;
}

export interface ClickCallbackProps {
  index: number;
}

/** Lifecycle callbacks */
export interface Callbacks {
  /** a callback called when a slide becomes active */
  view?: Callback<ViewCallbackProps>;
  /** a callback called when a slide is clicked */
  click?: Callback<ClickCallbackProps>;
  /** a callback called when the portal starts opening */
  entering?: Callback;
  /** a callback called when the portal opens */
  entered?: Callback;
  /** a callback called when the portal starts closing */
  exiting?: Callback;
  /** a callback called when the portal closes */
  exited?: Callback;
}

/** Custom UI labels / translations */
export interface Labels {
  Previous?: string;
  Next?: string;
  Close?: string;
}

export type Label = keyof Labels;

/** Toolbar settings */
export interface ToolbarSettings {
  /** buttons to render in the toolbar */
  buttons: (ToolbarButtonKey | React.ReactNode)[];
}
// ...
export type Plugin = (props: PluginProps) => void;

/** Deep partial utility type */
export type DeepPartial<
  T extends object,
  K extends keyof T = keyof T,
  E extends string = never,
> = Omit<Partial<T>, K> & {
  [P in K]?: DeepPartialValue<T[P], E>;
};

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export type DeepPartialValue<T, E extends string = never> = T extends any[]
  ? T
  : // eslint-disable-next-line @typescript-eslint/no-explicit-any
    T extends (...props: any[]) => any
    ? T
    : T extends object
      ? {
          [P in keyof T]?: P extends E ? T[P] : DeepPartialValue<T[P], E>;
        }
      : T;